Move JSON encode/decode fuzz tests to beacon/engine
to avoid import cycle and include all fuzz tests in a single place
